Add files via upload
Updated the while True to work with the updated/new building and buildingcalc classes.  Now asks the user to enter the building type, occupancy level, and time of day they would like to see the average energy for. The user can put in any combination of the possible options (4 building types, 3 occupancy levels, and 4 times of day) and the average total energy use will be displayed. Since the classes were fixed, the while True no longer needs the energy consumption function and was removed. Within the estimate function, new returns were added for both the energy minimum and maximum, and within the while True, lines were added to inform the user of the energy consumption range of the data for their three chosen parameters.