Add files via upload
Following changes: 
Identified the built in libraries/modules to achieve requirement
Pandas Dataframe & buildings dictionary are mutable. Strings & Tuples are immutable. Added tuple for TIMEOFDAY
Added in operator overloading with len() inside the Building Class. This allows us to call self instead of self.df. Minor change but fulfills criteria for operator overloading & makes code neater.
Created a new function row_check in buildingcalc class. This is a generator function which prints the rows that match the filter. This was previously done in a more roundabout manner. Method is neater & meets generator criteria.