Add files via upload
Plots updated to include displays of temperature, humidity, and energy. The plot including all of the values on top of one another is still there. Temperature values updated to display the Fahrenheit on the graph rather than the Celsius (to make it easier for viewers to understand the data seen). Some testing displays were also changed for clarity.