Add files via upload
Created three tests: 1 checks if the correct timeframe is assigned based on timeofday data. The other two check if the data columns of temp, humidity, and energy usage are all numbers (one passes and one fails for two separate cases to show functionality). These are housed in a test_.py file.